data/models/own/requirements.txt
pytorch-lightning==2.2.4
torchio==0.19.7
torchvision==0.17.2